separate out format module (pub(crate))
